+++ src/problems/vrptw/vrptw.h
@@ -23,7 +23,8 @@ public:
   explicit VRPTW(const Input& input);
 
   Solution
-  solve(unsigned exploration_level,
+  solve(unsigned nb_searches,
+        unsigned depth,
         unsigned nb_threads,
         const Timeout& timeout,
         const std::vector<HeuristicParameters>& h_param) const override;