+++ src/structures/vroom/input/input.cpp
@@ -33,11 +33,6 @@ Input::Input(io::Servers servers, ROUTER router, bool apply_TSPFix)
   : _apply_TSPFix(apply_TSPFix), _servers(std::move(servers)), _router(router) {
 }
 
-void Input::set_amount_size(unsigned amount_size) {
-  _amount_size = amount_size;
-  _zero = Amount(amount_size);
-}
-
 void Input::set_geometry(bool geometry) {
   _geometry = geometry;
 }
@@ -103,21 +98,27 @@ void Input::add_routing_wrapper(const std::string& profile) {
 #endif
 }
 
-void Input::check_job(Job& job) {
-  // Ensure delivery size consistency.
-  if (const auto delivery_size = job.delivery.size();
-      delivery_size != _amount_size) {
-    throw InputException(
-      "Inconsistent delivery length: " + std::to_string(delivery_size) +
-      " instead of " + std::to_string(_amount_size) + '.');
-  }
+void Input::check_amount_size(const Amount& amount) {
+  const auto size = amount.size();
 
-  // Ensure pickup size consistency.
-  if (const auto pickup_size = job.pickup.size(); pickup_size != _amount_size) {
-    throw InputException(
-      "Inconsistent pickup length: " + std::to_string(pickup_size) +
-      " instead of " + std::to_string(_amount_size) + '.');
+  if (!_amount_size.has_value()) {
+    // Only setup once on first call.
+    _amount_size = size;
+    _zero = Amount(size);
+  } else {
+    if (size != _amount_size.value()) {
+      throw InputException(
+        std::format("Inconsistent delivery length: {} instead of {}.",
+                    size,
+                    _amount_size.value()));
+    }
   }
+}
+
+void Input::check_job(Job& job) {
+  // Ensure delivery and pickup size consistency.
+  check_amount_size(job.delivery);
+  check_amount_size(job.pickup);
 
   // Ensure that location index are either always or never provided.
   bool has_location_index = job.location.user_index();
@@ -176,7 +177,7 @@ void Input::add_job(const Job& job) {
     throw InputException("Wrong job type.");
   }
   if (job_id_to_rank.contains(job.id)) {
-    throw InputException("Duplicate job id: " + std::to_string(job.id) + ".");
+    throw InputException(std::format("Duplicate job id: {}.", job.id));
   }
   job_id_to_rank[job.id] = jobs.size();
   jobs.push_back(job);
@@ -186,47 +187,50 @@ void Input::add_job(const Job& job) {
 
 void Input::add_shipment(const Job& pickup, const Job& delivery) {
   if (pickup.priority != delivery.priority) {
-    throw InputException("Inconsistent shipment priority for pickup " +
-                         std::to_string(pickup.id) + " and delivery " +
-                         std::to_string(delivery.id) + ".");
+    throw InputException(
+      std::
+        format("Inconsistent shipment priority for pickup {} and delivery {}.",
+               pickup.id,
+               delivery.id));
   }
   if (!(pickup.pickup == delivery.delivery)) {
-    throw InputException("Inconsistent shipment amount for pickup " +
-                         std::to_string(pickup.id) + " and delivery " +
-                         std::to_string(delivery.id) + ".");
+    throw InputException(
+      std::format("Inconsistent shipment amount for pickup {} and delivery {}.",
+                  pickup.id,
+                  delivery.id));
   }
   if (pickup.skills.size() != delivery.skills.size()) {
-    throw InputException("Inconsistent shipment skills for pickup " +
-                         std::to_string(pickup.id) + " and delivery " +
-                         std::to_string(delivery.id) + ".");
+    throw InputException(
+      std::format("Inconsistent shipment skills for pickup {} and delivery {}.",
+                  pickup.id,
+                  delivery.id));
   }
   for (const auto s : pickup.skills) {
     if (!delivery.skills.contains(s)) {
-      throw InputException("Inconsistent shipment skills for pickup " +
-                           std::to_string(pickup.id) + " and " +
-                           std::to_string(delivery.id) + ".");
+      throw InputException(
+        std::format("Inconsistent shipment skills for pickup {} and {}.",
+                    pickup.id,
+                    delivery.id));
     }
   }
 
   if (pickup.type != JOB_TYPE::PICKUP) {
-    throw InputException("Wrong type for pickup " + std::to_string(pickup.id) +
-                         ".");
+    throw InputException(std::format("Wrong type for pickup {}.", pickup.id));
   }
   if (pickup_id_to_rank.contains(pickup.id)) {
-    throw InputException("Duplicate pickup id: " + std::to_string(pickup.id) +
-                         ".");
+    throw InputException(std::format("Duplicate pickup id: {}.", pickup.id));
   }
   pickup_id_to_rank[pickup.id] = jobs.size();
   jobs.push_back(pickup);
   check_job(jobs.back());
 
   if (delivery.type != JOB_TYPE::DELIVERY) {
-    throw InputException("Wrong type for delivery " +
-                         std::to_string(delivery.id) + ".");
+    throw InputException(
+      std::format("Wrong type for delivery {}.", delivery.id));
   }
   if (delivery_id_to_rank.contains(delivery.id)) {
     throw InputException(
-      "Duplicate delivery id: " + std::to_string(delivery.id) + ".");
+      std::format("Duplicate delivery id: {}.", delivery.id));
   }
   delivery_id_to_rank[delivery.id] = jobs.size();
   jobs.push_back(delivery);
@@ -240,12 +244,7 @@ void Input::add_vehicle(const Vehicle& vehicle) {
   auto& current_v = vehicles.back();
 
   // Ensure amount size consistency.
-  if (const auto vehicle_amount_size = current_v.capacity.size();
-      vehicle_amount_size != _amount_size) {
-    throw InputException(
-      "Inconsistent capacity length: " + std::to_string(vehicle_amount_size) +
-      " instead of " + std::to_string(_amount_size) + '.');
-  }
+  check_amount_size(current_v.capacity);
 
   // Check for time-windows and skills.
   _has_TW = _has_TW || !vehicle.tw.is_default() || !vehicle.breaks.empty();
@@ -606,7 +605,8 @@ void Input::set_vehicles_costs() {
 }
 
 void Input::set_vehicles_max_tasks() {
-  if (_has_jobs && !_has_shipments && _amount_size > 0) {
+  if (const auto amount_size = get_amount_size();
+      _has_jobs && !_has_shipments && amount_size > 0) {
     // For job-only instances where capacity restrictions apply:
     // compute an upper bound of the number of jobs for each vehicle
     // based on pickups load and delivery loads. This requires sorting
@@ -621,12 +621,12 @@ void Input::set_vehicles_max_tasks() {
     };
 
     std::vector<std::vector<JobAmount>>
-      job_pickups_per_component(_amount_size,
+      job_pickups_per_component(amount_size,
                                 std::vector<JobAmount>(jobs.size()));
     std::vector<std::vector<JobAmount>>
-      job_deliveries_per_component(_amount_size,
+      job_deliveries_per_component(amount_size,
                                    std::vector<JobAmount>(jobs.size()));
-    for (std::size_t i = 0; i < _amount_size; ++i) {
+    for (std::size_t i = 0; i < amount_size; ++i) {
       for (Index j = 0; j < jobs.size(); ++j) {
         job_pickups_per_component[i][j] = JobAmount({j, jobs[j].pickup[i]});
         job_deliveries_per_component[i][j] =
@@ -643,7 +643,7 @@ void Input::set_vehicles_max_tasks() {
     for (Index v = 0; v < vehicles.size(); ++v) {
       std::size_t max_tasks = jobs.size();
 
-      for (std::size_t i = 0; i < _amount_size; ++i) {
+      for (std::size_t i = 0; i < amount_size; ++i) {
         Capacity pickup_sum = 0;
         Capacity delivery_sum = 0;
         std::size_t doable_pickups = 0;
@@ -783,9 +783,10 @@ void Input::set_vehicle_steps_ranks() {
       if (step.type == STEP_TYPE::BREAK) {
         auto search = current_vehicle.break_id_to_rank.find(step.id);
         if (search == current_vehicle.break_id_to_rank.end()) {
-          throw InputException("Invalid break id " + std::to_string(step.id) +
-                               " for vehicle " +
-                               std::to_string(current_vehicle.id) + ".");
+          throw InputException(
+            std::format("Invalid break id {} for vehicle {}.",
+                        step.id,
+                        current_vehicle.id));
         }
         step.rank = search->second;
       }
@@ -796,16 +797,18 @@ void Input::set_vehicle_steps_ranks() {
         case JOB_TYPE::SINGLE: {
           auto search = job_id_to_rank.find(step.id);
           if (search == job_id_to_rank.end()) {
-            throw InputException("Invalid job id " + std::to_string(step.id) +
-                                 " for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::format("Invalid job id {} for vehicle {}.",
+                          step.id,
+                          current_vehicle.id));
           }
           step.rank = search->second;
 
           if (planned_job_ids.contains(step.id)) {
-            throw InputException("Duplicate job id " + std::to_string(step.id) +
-                                 " in input steps for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::format("Duplicate job id {} in input steps for vehicle {}.",
+                          step.id,
+                          current_vehicle.id));
           }
           planned_job_ids.insert(step.id);
           break;
@@ -813,17 +816,19 @@ void Input::set_vehicle_steps_ranks() {
         case JOB_TYPE::PICKUP: {
           auto search = pickup_id_to_rank.find(step.id);
           if (search == pickup_id_to_rank.end()) {
-            throw InputException("Invalid pickup id " +
-                                 std::to_string(step.id) + " for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::format("Invalid pickup id {} for vehicle {}.",
+                          step.id,
+                          current_vehicle.id));
           }
           step.rank = search->second;
 
           if (planned_pickup_ids.contains(step.id)) {
-            throw InputException("Duplicate pickup id " +
-                                 std::to_string(step.id) +
-                                 " in input steps for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::
+                format("Duplicate pickup id {} in input steps for vehicle {}.",
+                       step.id,
+                       current_vehicle.id));
           }
           planned_pickup_ids.insert(step.id);
           break;
@@ -831,17 +836,19 @@ void Input::set_vehicle_steps_ranks() {
         case JOB_TYPE::DELIVERY: {
           auto search = delivery_id_to_rank.find(step.id);
           if (search == delivery_id_to_rank.end()) {
-            throw InputException("Invalid delivery id " +
-                                 std::to_string(step.id) + " for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::format("Invalid delivery id {} for vehicle {}.",
+                          step.id,
+                          current_vehicle.id));
           }
           step.rank = search->second;
 
           if (planned_delivery_ids.contains(step.id)) {
-            throw InputException("Duplicate delivery id " +
-                                 std::to_string(step.id) +
-                                 " in input steps for vehicle " +
-                                 std::to_string(current_vehicle.id) + ".");
+            throw InputException(
+              std::format("Duplicate delivery id {} in input steps for vehicle "
+                          "{}.",
+                          step.id,
+                          current_vehicle.id));
           }
           planned_delivery_ids.insert(step.id);
           break;
@@ -1088,7 +1095,8 @@ std::unique_ptr<VRP> Input::get_problem() const {
   return std::make_unique<CVRP>(*this);
 }
 
-Solution Input::solve(unsigned exploration_level,
+Solution Input::solve(unsigned nb_searches,
+                      unsigned depth,
                       unsigned nb_thread,
                       const Timeout& timeout,
                       const std::vector<HeuristicParameters>& h_param) {
@@ -1134,7 +1142,8 @@ Solution Input::solve(unsigned exploration_level,
   // Solve.
   const std::vector<HeuristicParameters> h_init_routes(1,
                                                        HEURISTIC::INIT_ROUTES);
-  auto sol = instance->solve(exploration_level,
+  auto sol = instance->solve(nb_searches,
+                             depth,
                              nb_thread,
                              solve_time,
                              _has_initial_routes ? h_init_routes : h_param);