+++ src/structures/vroom/input/input.h
@@ -12,6 +12,7 @@ All rights reserved (see LICENSE).
 
 #include <chrono>
 #include <memory>
+#include <optional>
 #include <unordered_map>
 
 #include "routing/wrapper.h"
@@ -82,14 +83,15 @@ private:
   // generating sparse matrices.
   std::unordered_map<Id, std::string> _vehicle_id_to_geometry;
 
-  unsigned _amount_size{0};
-  Amount _zero{0};
+  std::optional<unsigned> _amount_size;
+  Amount _zero;
 
   const io::Servers _servers;
   const ROUTER _router;
 
   std::unique_ptr<VRP> get_problem() const;
 
+  void check_amount_size(const Amount& amount);
   void check_job(Job& job);
 
   UserCost check_cost_bound(const Matrix<UserCost>& matrix) const;
@@ -123,10 +125,9 @@ public:
         ROUTER router = ROUTER::OSRM,
         bool apply_TSPFix = false);
 
-  void set_amount_size(unsigned amount_size);
-
   unsigned get_amount_size() const {
-    return _amount_size;
+    assert(_amount_size.has_value());
+    return _amount_size.value();
   }
 
   void set_geometry(bool geometry);
@@ -188,7 +189,8 @@ public:
   // Returns true iff both vehicles have common job candidates.
   bool vehicle_ok_with_vehicle(Index v1_index, Index v2_index) const;
 
-  Solution solve(unsigned exploration_level,
+  Solution solve(unsigned nb_searches,
+                 unsigned depth,
                  unsigned nb_thread,
                  const Timeout& timeout = Timeout(),
                  const std::vector<HeuristicParameters>& h_param =