+++ src/structures/vroom/job.cpp
@@ -10,7 +10,18 @@ All rights reserved (see LICENSE).
 #include "structures/vroom/job.h"
 #include "utils/helpers.h"
 
+#include <numeric>
+
 namespace vroom {
+
+inline Duration get_tw_length(const std::vector<TimeWindow>& tws) {
+  return std::accumulate(tws.begin(),
+                         tws.end(),
+                         0,
+                         [](auto sum, const auto& tw) {
+                           return sum + tw.length;
+                         });
+};
 
 Job::Job(Id id,
          const Location& location,
@@ -32,6 +43,7 @@ Job::Job(Id id,
     skills(std::move(skills)),
     priority(priority),
     tws(tws),
+    tw_length(get_tw_length(tws)),
     description(std::move(description)) {
   utils::check_tws(tws, id, "job");
   utils::check_priority(priority, id, "job");
@@ -57,6 +69,7 @@ Job::Job(Id id,
     skills(std::move(skills)),
     priority(priority),
     tws(tws),
+    tw_length(get_tw_length(tws)),
     description(std::move(description)) {
   assert(type == JOB_TYPE::PICKUP || type == JOB_TYPE::DELIVERY);
   utils::check_tws(tws, id, "job");