+++ src/structures/vroom/solution_indicators.h
@@ -18,7 +18,7 @@ All rights reserved (see LICENSE).
 
 namespace vroom::utils {
 
-template <class Route> struct SolutionIndicators {
+struct SolutionIndicators {
   Priority priority_sum{0};
   unsigned assigned{0};
   Eval eval;
@@ -26,6 +26,7 @@ template <class Route> struct SolutionIndicators {
 
   SolutionIndicators() = default;
 
+  template <class Route>
   SolutionIndicators(const Input& input, const std::vector<Route>& sol)
     : SolutionIndicators() {
     Index v_rank = 0;
@@ -56,6 +57,23 @@ template <class Route> struct SolutionIndicators {
                                                   rhs.eval.duration,
                                                   rhs.eval.distance);
   }
+
+#ifdef LOG_LS
+  friend bool operator==(const SolutionIndicators& lhs,
+                         const SolutionIndicators& rhs) {
+    return std::tie(rhs.priority_sum,
+                    rhs.assigned,
+                    lhs.eval.cost,
+                    lhs.used_vehicles,
+                    lhs.eval.duration,
+                    lhs.eval.distance) == std::tie(lhs.priority_sum,
+                                                   lhs.assigned,
+                                                   rhs.eval.cost,
+                                                   rhs.used_vehicles,
+                                                   rhs.eval.duration,
+                                                   rhs.eval.distance);
+  }
+#endif
 };
 
 } // namespace vroom::utils