+++ src/utils/helpers.cpp
@@ -63,27 +63,6 @@ SORT get_sort(std::string_view s) {
 }
 
 #ifdef LOG_LS_OPERATORS
-const std::array<std::string, OperatorName::MAX>
-  operator_names({"UnassignedExchange",
-                  "CrossExchange",
-                  "MixedExchange",
-                  "TwoOpt",
-                  "ReverseTwoOpt",
-                  "Relocate",
-                  "OrOpt",
-                  "IntraExchange",
-                  "IntraCrossExchange",
-                  "IntraMixedExchange",
-                  "IntraRelocate",
-                  "IntraOrOpt",
-                  "IntraTwoOpt",
-                  "PDShift",
-                  "RouteExchange",
-                  "SwapStar",
-                  "RouteSplit",
-                  "PriorityReplace",
-                  "TSPFix"});
-
 void log_LS_operators(
   const std::vector<std::array<ls::OperatorStats, OperatorName::MAX>>&
     ls_stats) {
@@ -108,7 +87,7 @@ void log_LS_operators(
   }
 
   for (auto op = 0; op < OperatorName::MAX; ++op) {
-    std::cout << operator_names[op] << "," << tried_sums[op] << ","
+    std::cout << OPERATOR_NAMES[op] << "," << tried_sums[op] << ","
               << applied_sums[op] << std::endl;
   }
   std::cout << "Total," << total_tried << "," << total_applied << std::endl;
@@ -226,15 +205,17 @@ void check_tws(const std::vector<TimeWindow>& tws,
                const Id id,
                const std::string& type) {
   if (tws.empty()) {
-    throw InputException("Empty time-windows for " + type + " " +
-                         std::to_string(id) + ".");
+    throw InputException(
+      std::format("Empty time-windows for {} {}.", type, id));
   }
 
   if (tws.size() > 1) {
     for (std::size_t i = 0; i < tws.size() - 1; ++i) {
       if (tws[i + 1].start <= tws[i].end) {
-        throw InputException("Unsorted or overlapping time-windows for " +
-                             type + " " + std::to_string(id) + ".");
+        throw InputException(
+          std::format("Unsorted or overlapping time-windows for {} {}.",
+                      type,
+                      id));
       }
     }
   }
@@ -244,9 +225,20 @@ void check_priority(const Priority priority,
                     const Id id,
                     const std::string& type) {
   if (priority > MAX_PRIORITY) {
-    throw InputException("Invalid priority value for " + type + " " +
-                         std::to_string(id) + ".");
+    throw InputException(
+      std::format("Invalid priority value for {} {}.", type, id));
   }
+}
+
+inline std::vector<Job> get_unassigned_jobs_from_ranks(
+  const Input& input,
+  const std::unordered_set<Index>& unassigned_ranks) {
+  std::vector<Job> unassigned_jobs;
+  std::ranges::transform(unassigned_ranks,
+                         std::back_inserter(unassigned_jobs),
+                         [&](auto j) { return input.jobs[j]; });
+
+  return unassigned_jobs;
 }
 
 Solution format_solution(const Input& input, const RawSolution& raw_routes) {
@@ -403,15 +395,9 @@ Solution format_solution(const Input& input, const RawSolution& raw_routes) {
                         v.description);
   }
 
-  // Handle unassigned jobs.
-  std::vector<Job> unassigned_jobs;
-  std::ranges::transform(unassigned_ranks,
-                         std::back_inserter(unassigned_jobs),
-                         [&](auto j) { return input.jobs[j]; });
-
   return Solution(input.zero_amount(),
                   std::move(routes),
-                  std::move(unassigned_jobs));
+                  get_unassigned_jobs_from_ranks(input, unassigned_ranks));
 }
 
 Route format_route(const Input& input,
@@ -910,15 +896,9 @@ Solution format_solution(const Input& input, const TWSolution& tw_routes) {
     }
   }
 
-  // Handle unassigned jobs.
-  std::vector<Job> unassigned_jobs;
-  std::ranges::transform(unassigned_ranks,
-                         std::back_inserter(unassigned_jobs),
-                         [&](auto j) { return input.jobs[j]; });
-
   return Solution(input.zero_amount(),
                   std::move(routes),
-                  std::move(unassigned_jobs));
+                  get_unassigned_jobs_from_ranks(input, unassigned_ranks));
 }
 
 } // namespace vroom::utils