+++ CHANGELOG.md
@@ -4,12 +4,20 @@
 
 ### Added
 
+#### Internals
+
+- LOG_LS flag to generate debug info on the internal solving process (#1124)
+
 ### Changed
 
 #### Internals
 
 - Bypass matrix request in `plan` mode (#444)
 - Refactor `Matrix` template class (#1089)
+- Refactor to use `std::format` whenever possible (#1081)
+- Reduce complexity for recreation process (#1155)
+- Refactor `SolutionIndicators` (#1169)
+- Remove amount consistency checks in `parse` in favor of upstream checks in `Input` (#1086)
 
 #### CI
 
@@ -18,8 +26,14 @@
 - Update Ubuntu image to 24.04 (#1080)
 - `vroom` workflow uses g++-14 and clang++-18 (#1080)
 - `vroom + libosrm` workflow uses g++-13 and clang++-17 (#1080)
+- Update clang-format to 18 (#1148)
 
 ### Fixed
+
+#### Core solving
+
+- Crash when input is valid JSON but not an object (#1172)
+- Capacity array check consistency (#1086)
 
 #### Internals
 
@@ -28,6 +42,10 @@
 #### CI
 
 - Wrong compiler used for clang-based OSRM builds (#1098)
+
+#### Routing
+
+- ORS error handling (#1083)
 
 ## [v1.14.0] - 2024-01-16